# pkg/api/rest/docs/swagger.yaml
basePath: /cicada
definitions:
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  model.CreateDataReq:
    properties:
      description:
        type: string
      task_groups:
        items:
          $ref: '#/definitions/model.CreateTaskGroupReq'
        type: array
    required:
    - task_groups
    type: object
...
  model.CreateTaskGroupReq:
    properties:
      description:
        type: string
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/model.CreateTaskReq'
        type: array
    required:
    - name
    - tasks
    type: object
  model.CreateTaskReq:
    properties:
      dependencies:
        items:
          type: string
        type: array
      extra:
        additionalProperties: true
        type: object
      name:
        type: string
      path_params:
        additionalProperties:
          type: string
        type: object
      query_params:
        additionalProperties:
          type: string
        type: object
      request_body:
        type: string
      task_component:
        type: string
    required:
    - name
    - request_body
    - task_component
    type: object
  model.CreateWorkflowReq:
    properties:
      data:
        $ref: '#/definitions/model.CreateDataReq'
      name:
        type: string
      spec_version:
        type: string
    required:
    - data
    - name
    type: object
  model.CreateWorkflowRunReq:
    properties:
      end_date:
        type: string
      execution_date:
        type: string
      run_type:
        type: string
      start_date:
        type: string
      state:
        type: string
      workflow_id:
        type: string
      workflow_run_id:
        type: string
    required:
    - workflow_id
    - workflow_run_id
    type: object
  model.Data:
    properties:
      description:
        type: string
      task_groups:
        items:
          $ref: '#/definitions/model.TaskGroup'
        type: array
    required:
    - task_groups
    type: object
...
  model.SimpleMsg:
    properties:
      message:
        type: string
    type: object
...
  model.Task:
    properties:
      dependencies:
        items:
          type: string
        type: array
      extra:
        additionalProperties: true
        type: object
      id:
        type: string
      name:
        type: string
      path_params:
        additionalProperties:
          type: string
        type: object
      query_params:
        additionalProperties:
          type: string
        type: object
      request_body:
        type: string
      task_component:
        type: string
      task_group_id:
        type: string
      workflow_id:
        type: string
    required:
    - id
    - name
    - request_body
    - task_component
    - task_group_id
    - workflow_id
    type: object
  model.TaskClearOption:
    properties:
      dryRun:
        type: boolean
      includeDownstream:
        type: boolean
      includeUpstream:
        description: |-
          IncludeFuture     bool     `json:"includeFuture"`
          IncludeParentdag  bool     `json:"includeParentdag"`
          IncludePast       bool     `json:"includePast"`
          IncludeSubdags    bool     `json:"includeSubdags"`
        type: boolean
      onlyFailed:
        type: boolean
      onlyRunning:
        type: boolean
      resetDagRuns:
        type: boolean
      taskIds:
        items:
          type: string
        type: array
    type: object
  model.TaskGroup:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      tasks:
        items:
          $ref: '#/definitions/model.Task'
        type: array
      workflow_id:
        type: string
    required:
    - id
    - name
    - tasks
    - workflow_id
    type: object
  model.TaskInstance:
    properties:
      duration_date:
        type: number
      end_date:
        type: string
      execution_date:
        type: string
      is_software_migration_task:
        type: boolean
      software_migration_execution_id:
        type: string
      start_date:
        type: string
      state:
        type: string
      task_id:
        type: string
      task_name:
        type: string
      try_number:
        type: integer
      workflow_id:
        type: string
      workflow_run_id:
        type: string
    type: object
  model.TaskInstanceReference:
    properties:
      execution_date:
        type: string
      task_id:
        description: The task ID.
        type: string
      task_name:
        type: string
      workflow_id:
        description: The DAG ID.
        type: string
      workflow_run_id:
        description: The DAG run ID.
        type: string
    type: object
...
  model.Workflow:
    properties:
      created_at:
        type: string
      data:
        $ref: '#/definitions/model.Data'
      id:
        type: string
      name:
        type: string
      spec_version:
        type: string
      updated_at:
        type: string
    required:
    - data
    - id
    - name
    - spec_version
    type: object
  model.WorkflowRun:
    properties:
      duration_date:
        type: number
      end_date:
        type: string
      execution_date:
        type: string
      id:
        type: string
      run_type:
        type: string
      start_date:
        type: string
      state:
        type: string
      workflow_id:
        type: string
    type: object
  model.WorkflowStatus:
    properties:
      count:
        type: integer
      state:
        type: string
    type: object
  model.WorkflowTemplate:
    properties:
      data:
        $ref: '#/definitions/model.Data'
      name:
        type: string
      spec_version:
        type: string
    required:
    - data
    - name
    - spec_version
    type: object
...
info:
  contact: {}
  description: Workflow management module
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: CM-Cicada REST API
  version: latest
paths:
...
  /workflow:
    get:
      consumes:
      - application/json
      description: Get a workflow list.
      operationId: list-workflow
      parameters:
      - description: Name of the workflow
        in: query
        name: name
        type: string
      - description: Page of the workflow list.
        in: query
        name: page
        type: string
      - description: Row of the workflow list.
        in: query
        name: row
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get a workflow list.
          schema:
            items:
              $ref: '#/definitions/model.Workflow'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get a workflow list.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: List Workflow
      tags:
      - '[Workflow]'
    post:
      consumes:
      - application/json
      description: Create a workflow.
      operationId: create-workflow
      parameters:
      - description: Workflow content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateWorkflowReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully create the workflow.
          schema:
            $ref: '#/definitions/model.WorkflowTemplate'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to create workflow.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create Workflow
      tags:
      - '[Workflow]'
  /workflow/{wfId}:
    delete:
      consumes:
      - application/json
      description: Delete the workflow.
      operationId: delete-workflow
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete the workflow
          schema:
            $ref: '#/definitions/model.SimpleMsg'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to delete the workflow
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Delete Workflow
      tags:
      - '[Workflow]'
    get:
      consumes:
      - application/json
      description: Get the workflow.
      operationId: get-workflow
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the workflow.
          schema:
            $ref: '#/definitions/model.Workflow'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the workflow.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Workflow
      tags:
      - '[Workflow]'
    put:
      consumes:
      - application/json
      description: Update the workflow content.
      operationId: update-workflow
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: Workflow to modify.
        in: body
        name: Workflow
        required: true
        schema:
          $ref: '#/definitions/model.CreateWorkflowReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update the workflow
          schema:
            $ref: '#/definitions/model.Workflow'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to update the workflow
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update Workflow
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/runs:
    get:
      consumes:
      - application/json
      description: Get DAG runs from Airflow API.
      operationId: get-workflow-runs
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the workflowRuns.
          schema:
            $ref: '#/definitions/model.WorkflowRun'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the workflowRuns.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get workflowRuns
      tags:
      - '[Workflow]'
    post:
      consumes:
      - application/json
      description: Airflow에서 DAG 종료 시 POST로 보낸 workflow run 데이터를 DB에 저장. 실행내역 전부 INSERT.
      operationId: create-workflow-run
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: Workflow run data (Airflow callback payload)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateWorkflowRunReq'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully saved workflow run
          schema:
            $ref: '#/definitions/model.WorkflowRun'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to save workflow run
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Create WorkflowRun (DB 저장)
      tags:
      - '[Workflow]'
  /workflow/{wfId}/status:
    get:
      consumes:
      - application/json
      description: Get the WorkflowStatus.
      operationId: get-WorkflowStatus
      parameters:
      - description: wfId of the workflow
        in: path
        name: wfId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the WorkflowVersion.
          schema:
            items:
              $ref: '#/definitions/model.WorkflowStatus'
            type: array
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the WorkflowVersion.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get WorkflowStatus
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/task/{taskId}:
    get:
      consumes:
      - application/json
      description: Get the task.
      operationId: get-task
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the task.
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the task.
          schema:
            $ref: '#/definitions/model.Task'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the task.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get Task
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/workflowRun/{wfRunId}/range:
    post:
      consumes:
      - application/json
      description: Clear the task Instance.
      operationId: clear-task-instances
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the wfRunId.
        in: path
        name: wfRunId
        required: true
        type: string
      - description: Workflow content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TaskClearOption'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully clear the taskInstances.
          schema:
            $ref: '#/definitions/model.TaskInstanceReference'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to clear the taskInstances.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Clear taskInstances
      tags:
      - '[Workflow]'
...
  /workflow/{wfId}/workflowRun/{wfRunId}/taskInstances:
    get:
      consumes:
      - application/json
      description: Get the task Logs.
      operationId: get-task-instances
      parameters:
      - description: ID of the workflow.
        in: path
        name: wfId
        required: true
        type: string
      - description: ID of the workflow.
        in: path
        name: wfRunId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get the taskInstances.
          schema:
            $ref: '#/definitions/model.TaskInstance'
        "400":
          description: Sent bad request.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Failed to get the taskInstances.
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get taskInstances
      tags:
      - '[Workflow]'
...
swagger: "2.0"